LPS-33869 Split CallbackPreferringTransactionInterceptor out to be more cleaner and efficient, don't backport.
