+++ portal-impl/src/META-INF/base-spring.xml
@@ -36,9 +36,10 @@
 	<bean id="indexableAdvice" class="com.liferay.portal.search.IndexableAdvice">
 		<property name="nextMethodInterceptor" ref="transactionAdvice" />
 	</bean>
-	<bean id="transactionAdvice" class="com.liferay.portal.spring.transaction.TransactionInterceptor">
+	<bean id="transactionInterceptorFactoryBean" class="com.liferay.portal.spring.transaction.TransactionInterceptorFactoryBean">
 		<property name="platformTransactionManager" ref="liferayTransactionManager" />
 		<property name="transactionAttributeSource" ref="transactionAttributeSource" />
 	</bean>
+	<bean id="transactionAdvice" factory-bean="transactionInterceptorFactoryBean" factory-method="create" />
 	<bean id="transactionAttributeSource" class="com.liferay.portal.spring.transaction.AnnotationTransactionAttributeSource" />
 </beans>
