+++ portal-impl/src/com/liferay/portal/spring/transaction/BaseTransactionInterceptor.java
@@ -1,0 +1,100 @@
+/**
+ * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.spring.transaction;
+
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+
+import java.lang.reflect.Method;
+
+import java.util.List;
+import java.util.concurrent.Callable;
+
+import org.aopalliance.intercept.MethodInterceptor;
+import org.aopalliance.intercept.MethodInvocation;
+
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.interceptor.TransactionAttribute;
+import org.springframework.transaction.interceptor.TransactionAttributeSource;
+
+/**
+ * @author Shuyang Zhou
+ */
+public abstract class BaseTransactionInterceptor implements MethodInterceptor {
+
+	public TransactionAttributeSource getTransactionAttributeSource() {
+		return transactionAttributeSource;
+	}
+
+	public Object invoke(MethodInvocation methodInvocation) throws Throwable {
+		Method method = methodInvocation.getMethod();
+
+		Class<?> targetClass = null;
+
+		Object targetBean = methodInvocation.getThis();
+
+		if (targetBean != null) {
+			targetClass = targetBean.getClass();
+		}
+
+		TransactionAttribute transactionAttribute =
+			transactionAttributeSource.getTransactionAttribute(
+				method, targetClass);
+
+		if (transactionAttribute == null) {
+			return methodInvocation.proceed();
+		}
+
+		return doInvoke(methodInvocation, transactionAttribute);
+	}
+
+	public void setPlatformTransactionManager(
+		PlatformTransactionManager platformTransactionManager) {
+
+		this.platformTransactionManager = platformTransactionManager;
+	}
+
+	public void setTransactionAttributeSource(
+		TransactionAttributeSource transactionAttributeSource) {
+
+		this.transactionAttributeSource = transactionAttributeSource;
+	}
+
+	protected abstract Object doInvoke(
+			MethodInvocation methodInvocation,
+			TransactionAttribute transactionAttribute)
+		throws Throwable;
+
+	protected void invokeCallbacks() {
+		List<Callable<?>> callables =
+			TransactionCommitCallbackUtil.popCallbackList();
+
+		for (Callable<?> callable : callables) {
+			try {
+				callable.call();
+			}
+			catch (Exception e) {
+				_log.error("Failed to execute transaction commit callback", e);
+			}
+		}
+	}
+
+	protected PlatformTransactionManager platformTransactionManager;
+	protected TransactionAttributeSource transactionAttributeSource;
+
+	private static Log _log = LogFactoryUtil.getLog(
+		BaseTransactionInterceptor.class);
+
+}
