+++ portal-impl/src/com/liferay/portal/spring/transaction/TransactionInterceptorFactoryBean.java
@@ -19,19 +19,41 @@ import com.liferay.portal.dao.jdbc.aop.DynamicDataSourceTransactionInterceptor;
 import com.liferay.portal.kernel.spring.util.FactoryBean;
 import com.liferay.portal.kernel.util.InfrastructureUtil;
 
+import org.aopalliance.intercept.MethodInterceptor;
+
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.interceptor.TransactionAttributeSource;
+
 /**
  * @author Shuyang Zhou
  */
 public class TransactionInterceptorFactoryBean
-	implements FactoryBean<TransactionInterceptor> {
+	implements FactoryBean<MethodInterceptor> {
 
-	public TransactionInterceptor create() {
+	public MethodInterceptor create() {
+		BaseTransactionInterceptor baseTransactionInterceptor;
+
+		if (_platformTransactionManager instanceof
+			CallbackPreferringTransactionInterceptor) {
+
+			baseTransactionInterceptor =
+				new CallbackPreferringTransactionInterceptor();
+		}
+		else {
+			baseTransactionInterceptor = new TransactionInterceptor();
+		}
+
+		baseTransactionInterceptor.setPlatformTransactionManager(
+			_platformTransactionManager);
+		baseTransactionInterceptor.setTransactionAttributeSource(
+			_transactionAttributeSource);
+
 		DynamicDataSourceTargetSource dynamicDataSourceTargetSource =
 			(DynamicDataSourceTargetSource)
 				InfrastructureUtil.getDynamicDataSourceTargetSource();
 
 		if (dynamicDataSourceTargetSource == null) {
-			return new TransactionInterceptor();
+			return baseTransactionInterceptor;
 		}
 
 		DynamicDataSourceTransactionInterceptor
@@ -40,8 +62,27 @@ public class TransactionInterceptorFactoryBean
 
 		dynamicDataSourceTransactionInterceptor.
 			setDynamicDataSourceTargetSource(dynamicDataSourceTargetSource);
+		dynamicDataSourceTransactionInterceptor.setTransactionAttributeSource(
+			_transactionAttributeSource);
+		dynamicDataSourceTransactionInterceptor.setTransactionInterceptor(
+			baseTransactionInterceptor);
 
 		return dynamicDataSourceTransactionInterceptor;
 	}
+
+	public void setPlatformTransactionManager(
+		PlatformTransactionManager platformTransactionManager) {
+
+		_platformTransactionManager = platformTransactionManager;
+	}
+
+	public void setTransactionAttributeSource(
+		TransactionAttributeSource transactionAttributeSource) {
+
+		_transactionAttributeSource = transactionAttributeSource;
+	}
+
+	private PlatformTransactionManager _platformTransactionManager;
+	private TransactionAttributeSource _transactionAttributeSource;
 
 }
