+++ portal-impl/src/com/liferay/portal/spring/util/SpringFactoryImpl.java
@@ -77,28 +77,26 @@ public class SpringFactoryImpl implements SpringFactory {
 		Object bean = InstanceFactory.newInstance(
 			ClassLoaderUtil.getPortalClassLoader(), className);
 
+		if (properties != null) {
+			for (Map.Entry<String, Object> entry : properties.entrySet()) {
+				String name = entry.getKey();
+
+				if (!allowedProperties.contains(name)) {
+					throw new SpringFactoryException(
+						"Undefined property " + name + " for class " +
+							className);
+				}
+
+				Object value = entry.getValue();
+
+				BeanPropertiesUtil.setProperty(bean, name, value);
+			}
+		}
+
 		if (bean instanceof FactoryBean) {
 			FactoryBean<Object> factoryBean = (FactoryBean<Object>)bean;
 
 			bean = factoryBean.create();
-		}
-
-		if (properties == null) {
-			return bean;
-		}
-
-		for (Map.Entry<String, Object> entry : properties.entrySet()) {
-			String name = entry.getKey();
-
-			if (!allowedProperties.contains(name)) {
-				throw new SpringFactoryException(
-					"Undefined property " + name + " for class " +
-						className);
-			}
-
-			Object value = entry.getValue();
-
-			BeanPropertiesUtil.setProperty(bean, name, value);
 		}
 
 		return bean;
